chore: rename worker scripts and keep deprecated aliases
