+++ .gitignore
@@ -85,7 +85,7 @@ skills/
 # Generated parser files
 src/core/parser/fbasic-parser.js
 
-# Built web worker (from scripts/build-service-worker.ts)
+# Built web worker (validated by scripts/dev/build-web-worker.ts)
 public/basic-interpreter-worker.js
 
 # Build number (auto-generated by vite plugin)
