+++ package.json
@@ -22,8 +22,9 @@
         "test:display-snapshots": "vitest run test/integration/SampleDisplaySnapshot.test.ts",
         "test:animation-integration": "vitest run test/integration/AnimationWorkerRuntimeIntegration.test.ts test/integration/AnimationSyncProtocol.test.ts test/integration/ClearAllMovements.test.ts test/integration/sprite-movement-lifecycle.test.ts test/integration/sprite-position-sync.test.ts",
         "test:coverage": "vitest run --coverage",
-        "build-web-worker": "tsx scripts/dev/build-service-worker.ts",
-        "build-service-worker": "pnpm build-web-worker",
+        "build-web-worker": "tsx scripts/dev/build-web-worker.ts",
+        "build-service-worker": "tsx scripts/dev/build-service-worker.ts",
+        "test-web-worker": "tsx scripts/dev/test-web-worker.ts",
         "test-service-worker": "tsx scripts/dev/test-service-worker.ts",
         "visualize-cst": "tsx scripts/dev/visualize-cst.ts",
         "check-syntax": "tsx scripts/validation/check-syntax.ts",
