+++ scripts/dev/build-service-worker.ts
@@ -1,21 +1,2 @@
-import { access } from 'node:fs/promises'
-import path from 'node:path'
-
-async function main(): Promise<void> {
-  const workerEntry = path.resolve(process.cwd(), 'src/core/workers/WebWorkerInterpreter.ts')
-
-  try {
-    await access(workerEntry)
-  } catch {
-    console.error(`[build-service-worker] Worker entrypoint not found: ${workerEntry}`)
-    process.exit(1)
-  }
-
-  console.log(`[build-service-worker] Worker entrypoint found: ${workerEntry}`)
-  console.log('[build-service-worker] Standalone worker build is not required; Vite bundles ?worker imports during app build.')
-}
-
-main().catch((error) => {
-  console.error('[build-service-worker] Unexpected error:', error)
-  process.exit(1)
-})
+console.warn('[build-service-worker] Deprecated alias. Use "pnpm build-web-worker" instead.')
+await import('./build-web-worker')
